public/requirement.txt
Pillow
pdf2image
python-docx
fpdf